fix cyclic imports to hopefully fix github ci
