+++ tests/unit/study/test_status.py
@@ -36,6 +36,7 @@ from datetime import datetime
 
 from deepdiff import DeepDiff
 
+from merlin.main import get_merlin_spec_with_override
 from merlin.study.status import Status
 from tests.unit.study.status_test_files import shared_tests, status_test_variables
 
@@ -68,6 +69,8 @@ class TestMerlinStatus(unittest.TestCase):
         """
         with self.assertRaises(ValueError):
             invalid_spec_path = f"{status_test_variables.PATH_TO_TEST_FILES}/nonexistent.yaml"
+            self.args.specification = invalid_spec_path
+            self.args.spec_provided, _ = get_merlin_spec_with_override(self.args)
             _ = Status(args=self.args, spec_display=True, file_or_ws=invalid_spec_path)
 
     def test_spec_setup_no_prompts(self):
@@ -78,6 +81,8 @@ class TestMerlinStatus(unittest.TestCase):
         correct. The methods covered here are _load_from_spec and _obtain_study,
         as well as any methods covered in assert_correct_attribute_creation
         """
+        self.args.specification = status_test_variables.SPEC_PATH
+        self.args.spec_provided, _ = get_merlin_spec_with_override(self.args)
         status_obj = Status(args=self.args, spec_display=True, file_or_ws=status_test_variables.SPEC_PATH)
         assert isinstance(status_obj, Status)
 
@@ -88,6 +93,10 @@ class TestMerlinStatus(unittest.TestCase):
         This is testing the prompt that's displayed when multiple study output
         directories are found. This tests the _obtain_study method using valid inputs.
         """
+        # We need to load in the MerlinSpec object and save it to the args we'll give to Status
+        self.args.specification = status_test_variables.SPEC_PATH
+        self.args.spec_provided, _ = get_merlin_spec_with_override(self.args)
+
         # We're going to load in a status object without prompts first and then use that to call the method
         # that prompts the user for input
         status_obj = Status(args=self.args, spec_display=True, file_or_ws=status_test_variables.SPEC_PATH)
@@ -98,6 +107,10 @@ class TestMerlinStatus(unittest.TestCase):
         This is testing the prompt that's displayed when multiple study output
         directories are found. This tests the _obtain_study method using invalid inputs.
         """
+        # We need to load in the MerlinSpec object and save it to the args we'll give to Status
+        self.args.specification = status_test_variables.SPEC_PATH
+        self.args.spec_provided, _ = get_merlin_spec_with_override(self.args)
+
         # We're going to load in a status object without prompts first and then use that to call the method
         # that prompts the user for input
         status_obj = Status(args=self.args, spec_display=True, file_or_ws=status_test_variables.SPEC_PATH)
